requirements.txt
aniso8601==9.0.1
black==23.1.0
click==8.1.3
Flask==2.2.3
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy==1.0.1
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
protobuf==3.20.3
pytz==2022.7.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
requests~=2.28.2